Rhino

+++ src/main/kotlin/com/chattriggers/ctjs/engine/module/ModulesGui.kt
@@ -28,20 +28,26 @@ object ModulesGui : GuiScreen() {
         var width = Renderer.screen.getWidth() - 100f
         if (width > 500) width = 500f
 
-        Renderer.drawRect(0x50000000, 0f, 0f, Renderer.screen.getWidth().toFloat(), Renderer.screen.getHeight().toFloat())
+        Renderer.drawRect(
+            0x50000000,
+            0f,
+            0f,
+            Renderer.screen.getWidth().toFloat(),
+            Renderer.screen.getHeight().toFloat()
+        )
 
         if (-window.scroll > window.height - Renderer.screen.getHeight() + 20)
             window.scroll = -window.height + Renderer.screen.getHeight() - 20
         if (-window.scroll < 0) window.scroll = 0f
 
         if (-window.scroll > 0) {
-            Renderer.drawRect(0xaa000000.toInt(), Renderer.screen.getWidth() - 20f, Renderer.screen.getHeight() - 20f, 20f, 20f)
+            Renderer.drawRect(0xaa000000, Renderer.screen.getWidth() - 20f, Renderer.screen.getHeight() - 20f, 20f, 20f)
             Renderer.drawString("^", Renderer.screen.getWidth() - 12f, Renderer.screen.getHeight() - 12f)
         }
 
         Renderer.drawRect(0x50000000, middle - width / 2f, window.scroll + 95f, width, window.height - 90)
 
-        Renderer.drawRect(0xaa000000.toInt(), middle - width / 2f, window.scroll + 95f, width, 25f)
+        Renderer.drawRect(0xaa000000, middle - width / 2f, window.scroll + 95f, width, 25f)
         window.title.draw(middle - width / 2f + 5, window.scroll + 100f)
         window.exit.setString(ChatLib.addColor("&cx")).draw(middle + width / 2f - 17, window.scroll + 99f)
 
@@ -65,7 +71,8 @@ object ModulesGui : GuiScreen() {
         }
 
         if (mouseX > Renderer.screen.getWidth() / 2f + width / 2f - 25 && mouseX < Renderer.screen.getWidth() / 2f + width / 2f
-        && mouseY > window.scroll + 95 && mouseY < window.scroll + 120) {
+            && mouseY > window.scroll + 95 && mouseY < window.scroll + 120
+        ) {
             Player.getPlayer()?.closeScreen()
             return
         }
